+++ host.ts
@@ -1,4 +1,5 @@
 import { editFolder } from "./folder-files";
+import { githubRemotes } from "./github-remote";
 import { experimental_defineHostEntry } from "@get-bb/plugin-sdk/host";
 import { moveHostContract } from "./move-contract";
 import { inspectMove, moveDirectory, linkDirectory } from "./move-files";
@@ -9,5 +10,6 @@ export default experimental_defineHostEntry({
     inspect: inspectMove,
     move: moveDirectory,
     link: linkDirectory,
+    github_remotes: githubRemotes,
   },
 });